Save the auth config
