+++ client/src/components/Error/Error.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { useRouteError } from 'react-router-dom';
+
+function Error() {
+  const error = useRouteError();
+
+  return (
+    <div id="error-page">
+      <h1>Oops!</h1>
+      <p>Sorry, an unexpected error has occurred.</p>
+      <p>
+        <i>{error.statusText || error.message}</i>
+      </p>
+    </div>
+  );
+}
+export default Error;
