+++ client/src/components/index.jsx
@@ -2,3 +2,4 @@ export { default as NavBar } from './Elements/navbar/NavBar';
 export { default as Home } from './Home/Home';
 export { default as Login } from './Login/Login';
 export { default as Register } from './Register/Register';
+export { default as Error } from './Error/Error';
