--- client/src/routes/Public.jsx
@@ -1,9 +1,0 @@
-import { Login } from '../components';
-
-// eslint-disable-next-line import/prefer-default-export
-export const publicRoutes = [
-  {
-    path: '/Login',
-    element: <Login />,
-  },
-];
